Fix JsProxy.to_py when default_converter is not provided

+++ src/core/jsproxy.c
@@ -719,8 +719,10 @@ JsProxy_toPy(PyObject* self,
   }
   PyObject* result =
     js2python_convert(GET_JSREF(self), depth, default_converter_js);
-  destroy_proxy(default_converter_js, NULL);
-  hiwire_CLEAR(default_converter_js);
+  if(default_converter_js != NULL){
+    destroy_proxy(default_converter_js, NULL);
+    hiwire_CLEAR(default_converter_js);
+  }
   return result;
 }
